Attempt start script
Add new start script to start frontend and api servers at the
same time.
Creates a copy of the frontend javascript file with set port
number and host address so the frontend knows where to query
for api calls.
Starts both the uvicorn fastapi server and the frontend
http server.
Overcomplicated it to heck initially with threading that
didnt work.